Special handling for IAR is problematic in zephyr
In zephyr, zephyr defined `__packed`. This is an IAR keyword.
Inside of zephyr, this is not problematic, since zephyr is not IAR aware
(yet). But since zephyr includes third-party modules such as mbedtls
which are, the redefinition of the `__packed` keyword causes compilation
problems becauise attribute((packed)) and __packed are not equivalent.
Since modern IAR defines both __ARM_FEATURE_UNALIGNED and
attribute((packed)) I think we can remove this special handling.
Modern IAR is required for zephyr anyways.

Signed-off-by: Robin Kastberg <robin.kastberg@iar.com>